refactor: rename FeedbackModal to Toast for better clarity, add minor updates to the component
