--- frontend/src/components/shared/modal/FeedbackModal/FeedbackModal.component.tsx
@@ -1,17 +1,0 @@
-import { ModalContainer } from './FeedbackModal.styles.ts';
-
-type ComponentPropsT = {
-  isVisible: boolean;
-  content: string;
-}
-
-const FeedbackModal = ({ isVisible, content }: ComponentPropsT) => {
-  return (
-    isVisible &&
-    <ModalContainer>
-      {content}
-    </ModalContainer>
-  );
-};
-
-export default FeedbackModal;
