--- frontend/src/components/shared/modal/FeedbackModal/index.ts
@@ -1,3 +1,0 @@
-import FeedbackModal from './FeedbackModal.component.tsx';
-
-export default FeedbackModal;
