+++ frontend/src/components/shared/notification/DefaultDialogStyles/DefaultDialogStyles.styles.ts
