+++ frontend/src/components/shared/notification/DefaultDialogStyles/index.ts
