+++ frontend/src/components/shared/notification/GlobalErrorModal/GlobalErrorModal.component.tsx
