+++ frontend/src/components/shared/notification/GlobalErrorModal/GlobalErrorModal.styles.ts
