+++ frontend/src/components/shared/notification/GlobalErrorModal/index.ts
