+++ frontend/src/components/shared/notification/GlobalLoadingModal/GlobalLoadingModal.styles.ts
