+++ frontend/src/components/shared/notification/Toast/index.ts
@@ -1,0 +1,3 @@
+import Toast from './Toast.component.tsx';
+
+export default Toast;
