+++ frontend/src/pages/shared/Home/Home.page.tsx
@@ -7,7 +7,7 @@ import {
   useRenderSelectedFormComponent,
   useShowConfirmationModal,
 } from './Home.hooks.tsx';
-import { ConfirmationModal } from '@components/shared/modal';
+import { ConfirmationModal } from 'components/shared/notification';
 import { MainContainer } from './Home.styles.ts';
 import { confirmationModalMessages } from './Home.utilities.ts';
