+++ frontend/src/pages/student/NewApplication/NewApplication.page.tsx
@@ -3,7 +3,7 @@ import {
   useGetCountryOptions,
   useGetUniversityOptionsByCountryUuid,
 } from '@hooks/index.ts';
-import { GlobalErrorModal } from '@components/shared/modal';
+import { GlobalErrorModal } from 'components/shared/notification';
 import { NewApplicationForm } from '@components/page/new-application';
 
 const NewApplicationPage = () => {
